requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
black==21.11b1
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
discord.py==1.7.3
idna==3.3
multidict==5.2.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
python-dotenv==0.19.2
regex==2021.11.10
requests==2.26.0
tomli==1.2.2
typing_extensions==4.0.0
urllib3==1.26.7
yarl==1.7.2
